requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
redis==5.0.4
pydantic==2.7.1
locust==2.25.0
prometheus-client==0.20.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-api==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-redis==0.46b0
httpx==0.27.0
aiohttp==3.9.5
python-dotenv==1.0.1
structlog==24.1.0
pytest==8.2.0
pytest-asyncio==0.23.7
